Move SetupIntegrationFormInputs to own file
